Inline method call
